Move delete flow runs and cancel-flow runs out of nav directory
