+++ src/components/SystemActions/ClearLate.vue
@@ -29,7 +29,7 @@ export default {
       if (!this.flowRuns?.length) return
 
       try {
-        const mutation = require('@/graphql/Nav/delete-flow-run.gql')
+        const mutation = require('@/graphql/Mutations/delete-flow-run.gql')
 
         this.loaded = 0
